Add ability to insert GIFs from keyboard
Closes vector-im/verticals-internal#21
