+++ changelog.d/8185.feature
@@ -1,0 +1,1 @@
+[Rich text editor] Add ability to insert GIFs from keyboard
